apis/apis: split up types.go
