+++ pkg/apis/apis/v1alpha1/types_apiresourceschema.go
@@ -1,0 +1,164 @@
+/*
+Copyright 2022 The KCP Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1alpha1
+
+import (
+	"encoding/json"
+
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
+// APIResourceSchema describes a resource, identified by (group, version, resource, schema).
+//
+// A APIResourceSchema is immutable and cannot be deleted if they are referenced by
+// an APIExport in the same workspace.
+//
+// +crd
+// +genclient
+// +genclient:nonNamespaced
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// +kubebuilder:resource:scope=Cluster,categories=kcp
+type APIResourceSchema struct {
+	metav1.TypeMeta `json:",inline"`
+	// +optional
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+
+	// Spec holds the desired state.
+	//
+	// +optional
+	Spec APIResourceSchemaSpec `json:"spec,omitempty"`
+}
+
+// APIResourceSchemaSpec defines the desired state of APIResourceSchema.
+type APIResourceSchemaSpec struct {
+	// group is the API group of the defined custom resource. Empty string means the
+	// core API group. 	The resources are served under `/apis/<group>/...` or `/api` for the core group.
+	//
+	// +required
+	Group string `json:"group"`
+
+	// names specify the resource and kind names for the custom resource.
+	//
+	// +required
+	Names apiextensionsv1.CustomResourceDefinitionNames `json:"names"`
+	// scope indicates whether the defined custom resource is cluster- or namespace-scoped.
+	// Allowed values are `Cluster` and `Namespaced`.
+	//
+	// +required
+	// +kubebuilder:validation:Enum=Cluster;Namespaced
+	Scope apiextensionsv1.ResourceScope `json:"scope"`
+
+	// versions is the API version of the defined custom resource.
+	//
+	// Note: the OpenAPI v3 schemas must be equal for all versions until CEL
+	//       version migration is supported.
+	//
+	// +required
+	// +listType=map
+	// +listMapKey=name
+	// +kubebuilder:validation:MinItems=1
+	Versions []APIResourceVersion `json:"versions"`
+}
+
+// APIResourceVersion describes one API version of a resource.
+type APIResourceVersion struct {
+	// name is the version name, e.g. “v1”, “v2beta1”, etc.
+	// The custom resources are served under this version at `/apis/<group>/<version>/...` if `served` is true.
+	//
+	// +required
+	// +kubebuilder:validation:MinLength=1
+	// +kubebuilder:validation:Pattern=^v[1-9][0-9]*([a-z]+[1-9][0-9]*)?$
+	Name string `json:"name"`
+	// served is a flag enabling/disabling this version from being served via REST APIs
+	//
+	// +required
+	// +kubebuilder:default=true
+	Served bool `json:"served"`
+	// storage indicates this version should be used when persisting custom resources to storage.
+	// There must be exactly one version with storage=true.
+	//
+	// +required
+	Storage bool `json:"storage"`
+	// deprecated indicates this version of the custom resource API is deprecated.
+	// When set to true, API requests to this version receive a warning header in the server response.
+	// Defaults to false.
+	//
+	// +optional
+	Deprecated bool `json:"deprecated,omitempty"`
+	// deprecationWarning overrides the default warning returned to API clients.
+	// May only be set when `deprecated` is true.
+	// The default warning indicates this version is deprecated and recommends use
+	// of the newest served version of equal or greater stability, if one exists.
+	//
+	// +optional
+	DeprecationWarning *string `json:"deprecationWarning,omitempty"`
+	// schema describes the structural schema used for validation, pruning, and defaulting
+	// of this version of the custom resource.
+	//
+	// +required
+	// +kubebuilder:pruning:PreserveUnknownFields
+	// +structType=atomic
+	Schema runtime.RawExtension `json:"schema"`
+	// subresources specify what subresources this version of the defined custom resource have.
+	//
+	// +optional
+	Subresources apiextensionsv1.CustomResourceSubresources `json:"subresources,omitempty"`
+	// additionalPrinterColumns specifies additional columns returned in Table output.
+	// See https://kubernetes.io/docs/reference/using-api/api-concepts/#receiving-resources-as-tables for details.
+	// If no columns are specified, a single column displaying the age of the custom resource is used.
+	//
+	// +optional
+	// +listType=map
+	// +listMapKey=name
+	AdditionalPrinterColumns []apiextensionsv1.CustomResourceColumnDefinition `json:"additionalPrinterColumns,omitempty"`
+}
+
+// APIResourceSchemaList is a list of APIResourceSchema resources
+//
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+type APIResourceSchemaList struct {
+	metav1.TypeMeta `json:",inline"`
+	metav1.ListMeta `json:"metadata"`
+
+	Items []APIResourceSchema `json:"items"`
+}
+
+func (v *APIResourceVersion) GetSchema() (*apiextensionsv1.JSONSchemaProps, error) {
+	if v.Schema.Raw == nil {
+		return nil, nil
+	}
+	var schema apiextensionsv1.JSONSchemaProps
+	if err := json.Unmarshal(v.Schema.Raw, &schema); err != nil {
+		return nil, err
+	}
+	return &schema, nil
+}
+
+func (v *APIResourceVersion) SetSchema(schema *apiextensionsv1.JSONSchemaProps) error {
+	if schema == nil {
+		v.Schema.Raw = nil
+		return nil
+	}
+	raw, err := json.Marshal(schema)
+	if err != nil {
+		return err
+	}
+	v.Schema.Raw = raw
+	return nil
+}
